feat: update README and package configurations
- Updated README.md to reflect project name change to "Art For Memories".
- Added next-sitemap dependency to package.json and package-lock.json.
- Configured next-sitemap in next-sitemap.config.js for sitemap generation.
- Modified postbuild script in package.json to include next-sitemap generation.
- Enhanced metadata for various pages to improve SEO.
- Refactored MemoryForm component to remove onSuccess prop.
- Updated Stories and VideoList components to streamline data fetching and rendering.
- Removed unnecessary client-side state management in several components.
