+++ next-sitemap.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  siteUrl: "https://artformemories.com",
+  generateRobotsTxt: true,
+};
