+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "postinstall": "prisma generate"
+    "postinstall": "prisma generate",
+    "postbuild": "next-sitemap"
   },
   "dependencies": {
     "@hookform/resolvers": "^4.1.3",
@@ -22,6 +23,7 @@
     "@tiptap/starter-kit": "^2.11.7",
     "axios": "^1.8.1",
     "cloudinary": "^2.6.0",
+    "next-sitemap": "^4.2.3",
     "datauri": "^4.1.0",
     "formidable": "^3.5.2",
     "framer-motion": "^12.4.7",
