+++ src/app/stories/illustrated/page.tsx
@@ -1,33 +1,16 @@
-"use client";
-
 import PageHeader from "@/components/headers/page-header";
 import Layout from "@/components/Layout";
-import StoriesList from "@/components/lists/StoriesList";
 import type { Stories } from "@/types/stories";
-import { useEffect, useState } from "react";
+import type { Metadata } from "next";
+import IllustratedStories from "./IllustratedStories";
+
+export const metadata: Metadata = {
+    title: "Illustrated Stories",
+    description: "Explore illustrated stories that bring memories to life, preserving the legacy of Genocide Survivors through art.",
+    keywords: ["Illustrated Stories", "Art of Memories Rwanda", "Genocide Survivors", "Preserve Memories", "Rwanda", "Art", "Genocide", "Survivors"],
+};
 
 export default function Stories() {
-    const [illustratedStories, setIllustratedStories] = useState([]);
-
-    const getAllStories = async () => {
-        try {
-            const response = await fetch('/api/stories');
-
-            if (response.ok) {
-                const allStories = await response.json();
-                setIllustratedStories(allStories.filter((story: Stories) => story.type === 'Illustrated'));
-            } else {
-                console.error('Failed to fetch stories');
-            }
-        } catch (error) {
-            console.error('Error fetching stories:', error);
-        }
-    }
-
-    useEffect(() => {
-        getAllStories();
-    })
-
     return (<>
         <Layout>
             <PageHeader
@@ -37,7 +20,7 @@ export default function Stories() {
             />
 
             <div className="bg-white">
-                <StoriesList title="Illustrated Stories" data={illustratedStories} />
+                <IllustratedStories />
             </div>
         </Layout>
     </>);
