+++ src/components/SubNav.tsx
@@ -19,7 +19,7 @@ export default function SubNav() {
                     </button>
 
                     <FormModal isOpen={isFormOpen} onClose={() => setFormOpen(false)}>
-                        <MemoryForm onSuccess={() => { }} />
+                        <MemoryForm />
                     </FormModal>
                 </div>
             </div>
@@ -66,7 +66,7 @@ export default function SubNav() {
                     </button>
 
                     <FormModal isOpen={isFormOpen} onClose={() => setFormOpen(false)}>
-                        <MemoryForm onSuccess={() => { }} />
+                        <MemoryForm />
                     </FormModal>
                 </div>
             </div>
